shader/decode: Split memory and texture instructions decoding
